OpenWrt router functionality

+++ wifiap.sh
@@ -1,6 +1,6 @@
 #!/bin/bash  
 
-#This script modify a variety of wireless network settings
+#This script modifies a variety of wireless network settings
 #
 # Usage: sudo sh wifiap.sh  [options]
 # 
@@ -10,22 +10,24 @@
 # -c,--channel                    Set channel number
 # -p,--password                   Set your passphrase (WiFi password)
 # -w,--wpa  [OFF/off]             Turn off wireless network security
-#
+# -r,--router                     Modifies the wireless network settings of OpenWrt router
 
 #set -x
 usage() { echo "Usage: sudo sh wifiap.sh  [options]" 
           echo " " 
           echo "[options]" 
 #          echo "-i,--interface [wlan0,wlan1]    Set interface name"
-          echo "-s,--ssid                       Set the name of your WiFi network"
-          echo "-c,--channel                    Set channel number"
-          echo "-p,--password                   Set your passphrase (WiFi password)"
-          echo "-w,--wpa  [OFF/off]             Turn off wireless network security" 1>&2; exit 1; }
+          echo "-s,--ssid                        Set the name of your WiFi network"
+          echo "-c,--channel                     Set channel number"
+          echo "-p,--password                    Set your passphrase (WiFi password)"
+          echo "-w,--wpa  [OFF/off]              Turn off wireless network security" 
+          echo "-r,--router                      Modifies the wireless network settings of OpenWrt router" 1>&2; exit 1; }
 
 
 
 path="/etc/hostapd/hostapd.conf"
-
+WRT="10.0.2.2"
+PSWD="mazi"
 ######  Parse command line arguments   ######
 
 while [ $# -gt 0 ]
@@ -33,10 +35,9 @@ do
 key="$1"
 
 case $key in
-#    -i |--interface)
-#    INTERFACE="$2"
-#    shift # past argument=value
-#    ;;
+    -r |--router)
+    ROUTER="TRUE"
+    ;;
     -s|--ssid)
     SSID="$2"
     shift # past argument=value
@@ -53,12 +54,8 @@ case $key in
     WPA="$2"
     shift # past argument=value
     ;;
-    --default)
-    DEFAULT=YES
-    shift # past argument with no value
-    ;;
     *)
-       # unknown option
+     # unknown option
     usage   
     ;;
 esac
@@ -77,33 +74,33 @@ fi
 
 ######  Setup arguments  ######
 
-#interface
-#if [ "$INTERFACE" ]; then
-#  
-#  sudo sed -i '/interface/d' $path
-#  sudo sed -i "$ a interface=$INTERFACE" $path  
-#fi
 
 #ssid
 if [ "$SSID" ]; then
-  
-  sudo sed -i '/ssid/d' $path
-  sudo sed -i "$ a ssid=$SSID" $path  
+  if [ "$ROUTER" ];then
+    sudo sshpass -p "$PSWD" ssh root@$WRT 'sed -i "/option ssid/c\        option ssid '$SSID'" /etc/config/wireless' 
+  else
+    sudo sed -i '/ssid/d' $path
+    sudo sed -i "$ a ssid=$SSID" $path
+  fi
 fi
 
 #channel
 if [ "$CHANNEL" ]; then
-  
-  sudo sed -i '/channel/d' $path
-  sudo sed -i "$ a channel=$CHANNEL" $path  
+  if [ "$ROUTER" ];then
+    sudo sshpass -p "$PSWD" ssh root@$WRT 'sed -i "/option channel/c\        option channel '$CHANNEL'" /etc/config/wireless'  
+  else
+    sudo sed -i '/channel/d' $path
+    sudo sed -i "$ a channel=$CHANNEL" $path  
+  fi
 fi
 
 #password
 if [ "$PASSWORD" ];then
-  if [ $(grep 'wpa' /etc/hostapd/hostapd.conf | wc -l) -eq 5  ];then  
+  if [ "$ROUTER" ];then  
     #Change password
-    sudo sed -i '/wpa_passphrase/d' $path
-    sudo sed -i "$ a wpa_passphrase=$PASSWORD" $path   
+    sudo sshpass -p "$PSWD" ssh root@$WRT 'sed -i "/option encryption/c\        option encryption 'psk2'" /etc/config/wireless'
+    sudo sshpass -p "$PSWD" ssh root@$WRT 'sed -i "/option key/c\        option key '$PASSWORD'" /etc/config/wireless'
   else
     #Set up wireless network security
     sudo sed -i '/wpa/d' $path
@@ -117,35 +114,38 @@ if [ "$PASSWORD" ];then
     sudo sed -i '$ a wpa_key_mgmt=WPA-PSK' $path
     sudo sed -i '$ a wpa_pairwise=TKIP CCMP' $path
     sudo sed -i '$ a wpa_ptk_rekey=600' $path
-
-  fi
+  fi  
 fi
 
 ######  Turn off wireless network security  #######
 
 if [ "$WPA" = "off" -o "$WPA" = "OFF" ];then
+  if [ "$ROUTER" ];then
+    sudo sshpass -p "$PSWD" ssh root@$WRT 'sed -i "/option encryption/c\        option encryption 'none'" /etc/config/wireless'
+  else
     sudo sed -i '/wpa/d' $path
     sudo sed -i '/wpa_passphrase/d' $path
     sudo sed -i '/wpa_key_mgmt/d' $path
     sudo sed -i '/wpa_pairwise/d' $path
     sudo sed -i '/wpa_ptk_rekey/d' $path
-
+  fi
 fi
 
 
 ###### Restart hostapd.conf ######
 
-id=$(ps aux | grep hostapd.conf| grep root| awk '{print $2}') 
 
-if [ "$id" ];then 
-   sudo kill $id
+if [ "$ROUTER" ];then
+  sudo sshpass -p "$PSWD" ssh root@$WRT 'uci set wireless.@wifi-device[0].disabled=1; uci commit wireless; wifi'
+  sudo sshpass -p "$PSWD" ssh root@$WRT 'uci set wireless.@wifi-device[0].disabled=0; uci commit wireless; wifi' 
+else 
+  id=$(ps aux | grep hostapd.conf| grep root| awk '{print $2}') 
+
+  if [ "$id" ];then 
+     sudo kill $id
+  fi
+  sleep 1
+  sudo ifconfig $(grep 'interface' /etc/hostapd/hostapd.conf| sed 's/\interface=//g') down
+  sudo hostapd -B $path
 fi
-sleep 1
-sudo ifconfig $(grep 'interface' /etc/hostapd/hostapd.conf| sed 's/\interface=//g') down
-sudo hostapd -B $path
-
 #set +x
-
-
-
-
